pyproject.toml
[tool.poetry]
name = "onshape_robotics_toolkit"
version = "0.0.1"
description = "A python library to facilitate interaction with Onshape's REST API"
authors = ["Senthur Ayyappan <senthurayyappan@gmail.com>"]
repository = "https://github.com/senthurayyappan/onshape-robotics-toolkit"
documentation = "https://senthurayyappan.github.io/onshape-robotics-toolkit/"
readme = "README.md"
packages = [
  {include = "onshape_robotics_toolkit"}
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
scipy = "^1.14.1"
numpy-stl = "^3.1.2"
networkx = "^3.4.2"
matplotlib = "^3.9.2"
pandas = "^2.2.3"
pyarrow = "^18.0.0"
regex = "^2024.9.11"
tqdm = "^4.67.0"
mujoco = "^3.2.6"
lxml = "^5.3.0"
optuna = "^4.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.16.2"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
pymdown-extensions = "^10.12"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = {extras = ["python"], version = "^0.26.1"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["onshape_robotics_toolkit"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true



[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # DoNotAssignLambda
    "E731",
    "TRY003",
    "SIM115",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["onshape_robotics_toolkit"]


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
